Updating dependencies, cleaning and adding ESLint
There is now an ESLint configuration to check
if the code follows the rules from the style guide.

There are also two new scripts in the package.json
for linting and formatting the entire codebase.

Keep in mind that you can't build the project if
you have style, TypeScript and/or other errors.